Update playwrightSettings.linux.json
Update .azure/pipelines/ci.yml


Update helix-matrix.yml

Update helix-matrix.yml

Break test to verify browser/platform coverage

Debug avail browsers

Update Build.props

Update ComponentTestBase.cs

Don't delete temp directory if not set

Update ComponentTestBase.cs

Update playwrightSettings.linux.json

Update playwrightSettings.ci.json

Update BinaryHttpClientTest.cs

Update BinaryHttpClientTest.cs

Update BinaryHttpClientTest.cs

Try turning off build in parallel

Create copy of basic test app -> mega test app

Update playwrightSettings.json

Fix publish path

Add new test project to slns, skip mariner

Revert chromium change on ci.linux

Reenable parallel builds

Turn off parallel again

Turn off test parallelization

Update src/Components/test/testassets/MegaTestApp/MegaTestApp.csproj


Delete mega test app

Update Microsoft.AspNetCore.Components.Migration.E2ETests.csproj

Update AspNetCore.sln

Update AspNetCore.sln

Update Build.props

Update Components.slnf

Update BrowserTestBase.cs

Update PlaywrightFixture.cs

+++ src/Components/test/E2ETestMigration/Tests/InputFileTest.cs
@@ -261,7 +261,10 @@ namespace Microsoft.AspNetCore.Components.E2ETest.Tests
         public override void Dispose()
         {
             base.Dispose();
-            Directory.Delete(_tempDirectory, recursive: true);
+            if (!String.IsNullOrEmpty(_tempDirectory))
+            {
+                Directory.Delete(_tempDirectory, recursive: true);
+            }
         }
 
         private struct TempFile
